packages/training/pyproject.toml
[project]
name = "ai-training-core"
version = "0.1.0"
description = "AI Training Core — RL/LLM training, market simulation, feature engineering"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.26",
    "pandas>=2.1",
    "torch>=2.1",
    "gymnasium>=0.29",
    "stable-baselines3>=2.2",
    "clickhouse-connect>=0.7",
    "qdrant-client>=1.7",
    "kafka-python>=2.0",
    "redis>=5.0",
    "websockets>=12.0",
    "ta>=0.11",
    "pydantic>=2.5",
    "requests>=2.31",
    "pyarrow>=14.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-asyncio>=0.23"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
